Move Extra.Map to -api
